fix: dedupe compareVersions after #323 moved it to core-version.ts
PR #323 moved `compareVersions` out of config-core.tsx into
`src/utils/core-version.ts` (hand-rolled). After merging main into
`fix/semver-dedup`, both files ended up with their own copy of
`compareVersions` — semver-based local + imported hand-rolled — which
tripped tsc (TS2440) and the lint chain (no-redeclare).

Make the shared `core-version.ts` semver-based (consistent with this
branch's intent) and drop the duplicate local definition + its `semver`
import in config-core.tsx. `useCoreBreakingConfirm` keeps working
unchanged because `isCoreBreakingVersion` still delegates to
`compareVersions`.
